# open-api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Agentrix API
  version: 2.0.0
  description: 'Canonical Agentrix API. Every route declares its authorization policy in `x-agentrix-policy`: anonymous callers hold the capabilities
    of the `spectator` role; `authenticated: true` additionally requires a signed-in user. Errors always use the `Error` envelope. This document
    is verified against the router by `src/server/openapi_contract_test.go`.'
servers:
- url: http://localhost:8080
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register a player account
      operationId: post_auth_register
      tags:
      - Sessions
      x-agentrix-policy: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /api/v1/auth/login:
    post:
      summary: Log in; sets the HttpOnly refresh cookie
      operationId: post_auth_login
      tags:
      - Sessions
      x-agentrix-policy: *id001
      description: Returns a short-lived access token (bearer). The refresh token is ONLY delivered as the HttpOnly cookie `agentrix_refresh`
        (Path=/api/v1/auth, SameSite=Strict, Secure in production); it never appears in the body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /api/v1/auth/refresh:
    post:
      summary: Rotate the refresh cookie and mint a new access token
      operationId: post_auth_refresh
      tags:
      - Sessions
      x-agentrix-policy: *id001
      description: 'Single-use rotation in one transaction. Reusing a rotated token revokes the whole session family (`refresh_token_reused`).
        Requires the CSRF header `X-Requested-With: agentrix`.'
      parameters:
      - name: X-Requested-With
        in: header
        required: true
        schema:
          type: string
          enum:
          - agentrix
  /api/v1/auth/logout:
    post:
      summary: Revoke the current session family
      operationId: post_auth_logout
      tags:
      - Sessions
      x-agentrix-policy: *id001
      description: Revokes the family identified by the refresh cookie (or the bearer session). Access tokens of that family stop working immediately.
      parameters:
      - name: X-Requested-With
        in: header
        required: true
        schema:
          type: string
          enum:
          - agentrix
  /api/v1/auth/logout-all:
    post:
      summary: Revoke every session of the current user
      operationId: post_auth_logout_all
      tags:
      - Sessions
      x-agentrix-policy:
        capabilities:
        - users:read:own
        authenticated: true
      security:
      - bearer: []
  /api/v1/me:
    get:
      summary: Current account with roles and effective capabilities
      operationId: get_me
      tags:
      - Users
      x-agentrix-policy:
        capabilities:
        - users:read:own
        authenticated: true
      security:
      - bearer: []
  /api/v1/users:
    get:
      summary: List accounts
      operationId: get_users
      tags:
      - Users
      x-agentrix-policy:
        capabilities:
        - users:read:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/UserStatus'
    post:
      summary: Create an account with explicit roles
      operationId: post_users
      tags:
      - Users
      x-agentrix-policy:
        capabilities:
        - users:roles:manage
        authenticated: true
      security:
      - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
  /api/v1/users/{id}:
    get:
      summary: Read an account (own or any)
      operationId: get_users_id
      tags:
      - Users
      x-agentrix-policy:
        capabilities:
        - users:read:own
        - users:read:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      summary: Update email or password
      operationId: patch_users_id
      tags:
      - Users
      x-agentrix-policy:
        capabilities:
        - users:update:own
        - users:update:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
  /api/v1/users/{id}/roles:
    put:
      summary: Replace the exact role set (effective on the next request)
      operationId: put_users_id_roles
      tags:
      - Users
      x-agentrix-policy:
        capabilities:
        - users:roles:manage
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolesRequest'
  /api/v1/users/{id}/status:
    put:
      summary: Set account status explicitly; leaving active revokes sessions
      operationId: put_users_id_status
      tags:
      - Users
      x-agentrix-policy:
        capabilities:
        - users:update:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserStatusRequest'
  /api/v1/games:
    get:
      summary: Installed game modules
      operationId: get_games
      tags:
      - Games
      x-agentrix-policy:
        capabilities:
        - contests:view
        authenticated: false
      security:
      - bearer: []
      - {}
  /api/v1/games/{id}:
    get:
      summary: Game module
      operationId: get_games_id
      tags:
      - Games
      x-agentrix-policy:
        capabilities:
        - contests:view
        authenticated: false
      security:
      - bearer: []
      - {}
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/v1/agents:
    get:
      summary: List own agents (owner=* or owner id requires agents:read:any)
      operationId: get_agents
      tags:
      - Agents
      x-agentrix-policy:
        capabilities:
        - agents:read:own
        - agents:read:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: owner
        in: query
        schema:
          type: string
    post:
      summary: Create an agent
      operationId: post_agents
      tags:
      - Agents
      x-agentrix-policy:
        capabilities:
        - agents:create
        authenticated: true
      security:
      - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
  /api/v1/agents/{id}:
    get:
      summary: Read an agent (own or any)
      operationId: get_agents_id
      tags:
      - Agents
      x-agentrix-policy:
        capabilities:
        - agents:read:own
        - agents:read:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      summary: Rename or describe an agent
      operationId: patch_agents_id
      tags:
      - Agents
      x-agentrix-policy:
        capabilities:
        - agents:update:own
        - agents:update:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
  /api/v1/agents/{id}/status:
    put:
      summary: Activate or disable an agent (explicit status)
      operationId: put_agents_id_status
      tags:
      - Agents
      x-agentrix-policy:
        capabilities:
        - agents:update:own
        - agents:update:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentStatusRequest'
  /api/v1/agents/{id}/submissions:
    get:
      summary: Submission versions of an agent
      operationId: get_agents_id_submissions
      tags:
      - Submissions
      x-agentrix-policy:
        capabilities:
        - submissions:read:own
        - submissions:read:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: 'Upload a bot bundle (ZIP: agentrix.json + bot.py); admission runs in the worker sandbox'
      operationId: post_agents_id_submissions
      tags:
      - Submissions
      x-agentrix-policy:
        capabilities:
        - submissions:create:own
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                bundle:
                  type: string
                  format: binary
                  description: ZIP <= 2 MiB with exactly agentrix.json and bot.py
              additionalProperties: false
              required:
              - bundle
  /api/v1/submissions/{id}:
    get:
      summary: Read a submission (status, digest, admission error)
      operationId: get_submissions_id
      tags:
      - Submissions
      x-agentrix-policy:
        capabilities:
        - submissions:read:own
        - submissions:read:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/v1/submissions/{id}/disable:
    post:
      summary: Retire a ready submission
      operationId: post_submissions_id_disable
      tags:
      - Submissions
      x-agentrix-policy:
        capabilities:
        - agents:update:own
        - agents:update:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/v1/contests:
    get:
      summary: List contests (drafts only with contests:manage)
      operationId: get_contests
      tags:
      - Contests
      x-agentrix-policy:
        capabilities:
        - contests:view
        authenticated: false
      security:
      - bearer: []
      - {}
    post:
      summary: Create a draft contest
      operationId: post_contests
      tags:
      - Contests
      x-agentrix-policy:
        capabilities:
        - contests:manage
        authenticated: true
      security:
      - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContestRequest'
  /api/v1/contests/{id}:
    get:
      summary: Read a contest
      operationId: get_contests_id
      tags:
      - Contests
      x-agentrix-policy:
        capabilities:
        - contests:view
        authenticated: false
      security:
      - bearer: []
      - {}
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      summary: Edit contest details (scoring policy only while draft)
      operationId: patch_contests_id
      tags:
      - Contests
      x-agentrix-policy:
        capabilities:
        - contests:manage
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContestRequest'
  /api/v1/contests/{id}/transitions:
    post:
      summary: Apply a state machine transition
      operationId: post_contests_id_transitions
      tags:
      - Contests
      x-agentrix-policy:
        capabilities:
        - contests:manage
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionRequest'
  /api/v1/contests/{id}/entries:
    get:
      summary: Contest roster
      operationId: get_contests_id_entries
      tags:
      - Entries
      x-agentrix-policy:
        capabilities:
        - contests:view
        authenticated: false
      security:
      - bearer: []
      - {}
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Enroll an agent locked to an exact ready submission
      operationId: post_contests_id_entries
      tags:
      - Entries
      x-agentrix-policy:
        capabilities:
        - entries:create:own
        - entries:manage:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollRequest'
  /api/v1/contests/{id}/entries/{entryId}/submission:
    put:
      summary: Re-lock the entry to another ready submission (registration open)
      operationId: put_contests_id_entries_entryId_submission
      tags:
      - Entries
      x-agentrix-policy:
        capabilities:
        - entries:create:own
        - entries:manage:any
        authenticated: true
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: entryId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResubmitRequest'
...
  /api/v1/contests/{id}/rankings:
    get:
      summary: Current ranking projection (committed runs only)
      operationId: get_contests_id_rankings
      tags:
      - Rankings
      x-agentrix-policy:
        capabilities:
        - rankings:view
        authenticated: false
      security:
      - bearer: []
      - {}
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT (HS256, 10 min, bound to a session family)
  schemas:
    UserStatus:
      type: string
      enum:
      - active
      - suspended
      - disabled
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          pattern: ^[A-Za-z0-9_.-]{3,32}$
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          maxLength: 128
      additionalProperties: false
      required:
      - username
      - email
      - password
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      additionalProperties: false
      required:
      - username
      - password
...
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
          minItems: 1
      additionalProperties: false
      required:
      - username
      - email
      - password
      - roles
...